python/requirements.txt
attrs==17.4.0
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
PyHamcrest==1.9.0
pytest==3.5.0
six==1.11.0
